+++ src/features/post/ui/PostItem.tsx
@@ -1,14 +1,17 @@
 import { PostHeader } from "./PostHeader";
 import { Table, TableBody, TableHeader } from "../../../shared/ui";
 import { PostRow } from "./PostRow";
-import { loadingAtom, postsAtom } from "../../../app/store/atom";
+import { postsAtom } from "../../../app/store/atom";
 import { useAtomValue } from "jotai";
+import { usePostsQuery } from "../../../entities/post/hook/usePostsQuery";
 
 export const PostItem: React.FC = () => {
   const posts = useAtomValue(postsAtom);
-  const loading = useAtomValue(loadingAtom);
 
-  if (loading) return <div className="flex justify-center p-4">로딩 중...</div>;
+  const { isLoading } = usePostsQuery();
+
+  if (isLoading)
+    return <div className="flex justify-center p-4">로딩 중...</div>;
 
   return (
     <Table>
