+++ src/pages/PostsManagerPage.tsx
@@ -1,20 +1,12 @@
-import { useEffect } from "react";
 import { Card, CardContent, CardHeader } from "../shared/ui";
 
 import { UserModal } from "../entities/user/ui/UserModal.tsx";
 import { DialogAddPost } from "../entities/post/ui/DialogAddPost.tsx";
 
-import { useAtomValue } from "jotai";
-import {
-  selectedTagAtom,
-  limitAtom,
-  skipAtom,
-  sortByAtom,
-  sortOrderAtom,
-} from "../app/store/atom.ts";
-import { usePosts } from "../entities/post/hook/usePosts.ts";
 import { useParams } from "../shared/hook/useParams.ts";
 import { useInitializePosts } from "../shared/hook/useInitializePosts.ts";
+import { usePostsQuery } from "../entities/post/hook/usePostsQuery.ts";
+
 import { DialogPostDetail } from "../entities/post/ui/DialogPostDetail.tsx";
 import { DialogEditPost } from "../entities/post/ui/DialogEditPost.tsx";
 import { DialogAddComment } from "../entities/comment/ui/DialogAddComment.tsx";
@@ -23,37 +15,14 @@ import { CardContentBody } from "../entities/card/ui/CardContentBody.tsx";
 import { CardTitleContents } from "../entities/card/ui/CardTitleContents.tsx";
 
 const PostsManager = () => {
-  // 전역 변수
-  const selectedTag = useAtomValue(selectedTagAtom);
-  const skip = useAtomValue(skipAtom);
-  const limit = useAtomValue(limitAtom);
-  const sortBy = useAtomValue(sortByAtom);
-  const sortOrder = useAtomValue(sortOrderAtom);
-
-  const { handleFetchPost, handleFetchPostsByTag } = usePosts();
+  useInitializePosts();
 
   const { updateURL } = useParams();
+  const { data: posts } = usePostsQuery();
 
-  // usePost
-  useEffect(() => {
-    if (selectedTag) {
-      handleFetchPostsByTag(selectedTag);
-    } else {
-      handleFetchPost();
-    }
+  if (posts) {
     updateURL();
-  }, [
-    skip,
-    limit,
-    sortBy,
-    sortOrder,
-    selectedTag,
-    handleFetchPost,
-    handleFetchPostsByTag,
-    updateURL,
-  ]);
-
-  useInitializePosts();
+  }
 
   return (
     <Card className="w-full max-w-6xl mx-auto">
